requirements.txt
aiofiles==0.4.0
aiohttp==3.4.4
aioredis==1.2.0
async-timeout==3.0.1
asyncio==3.4.3
attrs==18.2.0
beautifulsoup4==4.6.0
browsercookie2==0.7.1
certifi==2018.4.16
chardet==3.0.4
cssselect==1.0.3
Django==2.0.1
entrypoints==0.2.3
fake-useragent==0.1.11
hiredis==0.3.0
httpsproxy-urllib2==1.0
idna==2.6
idna-ssl==1.1.0
keyring==13.1.0
lxml==4.2.1
M2CryptoWin64==0.21.1.post3
motor==2.0.0
multidict==4.5.1
mysql-connector==2.1.6
numpy==1.15.4
pandas==0.23.4
pycryptodome==3.6.3
pymongo==3.7.2
pypiwin32==223
PyQt5==5.11.3
PyQt5-sip==4.19.13
python-dateutil==2.7.5
pytz==2017.3
pywin32==223
pywin32-ctypes==0.1.2
readability==0.3
readability-lxml==0.7
redis==3.0.1
requests==2.18.4
retrying==1.3.3
six==1.11.0
ujson==1.35
urllib3==1.22
useragent==0.1.1
yarl==1.2.6
